Pull up following revision(s) (requested by kre in ticket #1139):
	sbin/dump/dump.h: revision 1.59
	sbin/dump/main.c: revision 1.78

PR bin/55834

count blocks written in unsigned 64 bit counter
rather than signed int which overflows after 2^31-1
blocks (2TiB) after which neither the 5 minute
status updates or SIGINFO (^T) reports are issued
until the negative numbers increase past 0 and
wildly inaccurate reports would be written.

+++ sbin/dump/dump.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: dump.h,v 1.57 2019/03/25 02:13:01 manu Exp $	*/
+/*	$NetBSD: dump.h,v 1.57.2.1 2020/12/06 10:31:02 martin Exp $	*/
 
 /*-
  * Copyright (c) 1980, 1993
@@ -134,7 +134,7 @@ int	unlimited;	/* if set, write to end of medium */
 extern int	density;	/* density in 0.1" units */
 extern int	notify;		/* notify operator flag */
 extern int	timestamp;	/* timestamp messages */
-extern int	blockswritten;	/* number of blocks written on current tape */
+extern u_int64_t	blockswritten;	/* blocks written on current tape */
 extern int	tapeno;		/* current tape number */
 extern int	is_ufs2;
